feat: split the project into two modules - core & app
