+++ .gitignore
@@ -27,3 +27,6 @@ bin/
 
 ### Mac OS ###
 .DS_Store
+
+### IDEA ###
+.idea
