+++ core/src/main/java/com/employeemanagement/core/Person.java
@@ -1,0 +1,23 @@
+package com.employeemanagement;
+
+public class Person {
+    private final String name;
+    private final int age;
+    private final Address address;
+
+    public Person(String name, int age, Address address) {
+        this.name = name;
+        this.age = age;
+        this.address = address;
+    }
+
+    public String getName() {
+        return name;
+    }
+    public int getAge() {
+        return age;
+    }
+    public String getAddress() {
+        return address.getAddressDetails();
+    }
+}
